// src/app/shared/models/database-model/productCampania.model.ts
export interface ProductoCampanaModel {
    IdProducto: string;
    SubFamilia: string;
    FamSubFamProducto: string;
    Producto: string;
    IdCampana: string;
    Activo: string;
    FamiliaProducto: string;
    ReporteVisibilidad: string;
    IdSubFamiliaProducto: string;
    IdLineaNegocio: string;
    UnidadNegocio: string;
}